# openapi.yaml
openapi: 3.0.3
info:
  title: Stellar-Spend API
  description: |
    Stellar-Spend converts Stellar USDC to fiat currency (off-ramp) using the
    Allbridge cross-chain bridge and Paycrest fiat settlement layer.

    ## Authentication

    | Route group | Auth method |
    |---|---|
    | `/api/offramp/*` (legacy) | None — server-side secrets only |
    | `/api/v1/*` | API key via `X-API-Key` or `Authorization: Bearer` |
    | `/api/api-keys/*` | Admin bearer token via `Authorization: Bearer <API_KEY_ADMIN_TOKEN>` |
    | `/api/auth/2fa/*` | Public — user-supplied credentials in body |
    | `/api/webhooks/paycrest` | HMAC-SHA256 via `X-Paycrest-Signature` |

    ## API Key Scopes

    | Scope | Endpoints |
    |---|---|
    | `offramp:read` | GET currencies, institutions, rate, bridge status |
    | `offramp:write` | POST quote, build-tx, submit-soroban, paycrest order |
    | `webhook:read` | Receive webhook events |

    ## Rate Limits

    `/api/v1/*` endpoints are rate-limited per API key.

    Default: **120 requests / 60 seconds**. The following headers are returned on
    every response:

    | Header | Description |
    |---|---|
    | `X-RateLimit-Limit` | Maximum requests in the current window |
    | `X-RateLimit-Remaining` | Requests remaining in the current window |
    | `X-RateLimit-Reset` | Unix timestamp when the window resets |
    | `Retry-After` | Seconds to wait (only on `429` responses) |

...

    ## Versioning

...

    Version can also be negotiated with:
...

...

...
servers:
...
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: Programmatic API key for `/api/v1/*` endpoints
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        For `/api/v1/*` endpoints use a programmatic API key.
        For `/api/api-keys/*` admin endpoints use `API_KEY_ADMIN_TOKEN`.
    PaycrestSignature:
      type: apiKey
      in: header
      name: X-Paycrest-Signature
      description: HMAC-SHA256 hex digest of the raw request body (webhook only)

  # ───────── Reusable response headers ─────────
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Machine-readable error code, stable across releases. Safe to branch client logic on.
          enum:
            - validation_error
            - not_found
            - unauthorized
            - forbidden
            - conflict
            - rate_limit_exceeded
            - server_error
            - external_service_error
          example: validation_error
        message:
          type: string
          description: Human-readable description of the error, safe to display to end users.
          example: Invalid request parameters
        details:
          type: object
          description: >
            Additional machine-readable context (e.g. per-field validation failures).
            Only populated outside production; omitted entirely when production hides internals.
          additionalProperties: true
...
    Currency:
      type: object
      properties:
        code:
          type: string
          example: NGN
        name:
          type: string
          example: Nigerian Naira
        symbol:
          type: string
          example: "₦"
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
          example: uuid
        name:
          type: string
          example: Production Partner
        keyPrefix:
          type: string
          example: abc123def456
        status:
          type: string
          enum: [active, revoked, rotated]
          example: active
        rateLimitMaxRequests:
          type: integer
          example: 120
        rateLimitWindowMs:
          type: integer
          example: 60000
    CreateApiKeyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: Production Partner
        rateLimitMaxRequests:
          type: integer
          example: 120
        rateLimitWindowMs:
          type: integer
          example: 60000
    CreateApiKeyResponse:
      type: object
      properties:
        data:
          allOf:
            - $ref: "#/components/schemas/ApiKey"
            - type: object
              properties:
                plaintextKey:
                  type: string
                  example: ssp_live_abc123def456.very-secret-value
...
paths:
...
  /api/offramp/currencies:
    get:
      tags: [offramp]
      summary: List supported fiat currencies
      operationId: getCurrencies
      description: Returns supported currencies. Response is cached for 5 minutes.
      responses:
        "200":
          description: List of currencies
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Currency"
              example:
                data:
                  - code: NGN
                    name: Nigerian Naira
                    symbol: "₦"
                  - code: KES
                    name: Kenyan Shilling
                    symbol: KSh
        "500":
          description: Paycrest API unreachable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/offramp/currencies:
    get:
      tags: [offramp]
      summary: List supported fiat currencies (v1)
      operationId: getCurrenciesV1
      security:
        - ApiKeyHeader: []
        - BearerAuth: []
      responses:
        "200":
          description: List of currencies
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Currency"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Paycrest API unreachable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Institutions ─────────────────────────────────────────────────
...
  /api/api-keys:
    get:
      tags: [api-keys]
      summary: List all API keys
      operationId: listApiKeys
      security:
        - BearerAuth: []
      description: Requires `API_KEY_ADMIN_TOKEN` as bearer token.
      responses:
        "200":
          description: List of API keys (no plaintext secrets)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApiKey"
        "401":
          description: Invalid or missing admin token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags: [api-keys]
      summary: Create a new API key
      operationId: createApiKey
      security:
        - BearerAuth: []
      description: |
        Requires `API_KEY_ADMIN_TOKEN` as bearer token.
        The plaintext key is returned **once** and cannot be retrieved again.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyRequest"
      responses:
        "201":
          description: API key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateApiKeyResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/auth/2fa/enforcement:
    get:
      tags: [auth]
      summary: Get the org-wide 2FA enforcement policy (admin)
      operationId: getEnforcementPolicy
      parameters:
        - name: x-admin-id
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current enforcement policy
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Missing x-admin-id header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags: [auth]
      summary: Update the org-wide 2FA enforcement policy (admin)
      operationId: updateEnforcementPolicy
      parameters:
        - name: x-admin-id
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Updated enforcement policy
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Missing x-admin-id header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags: [auth]
      summary: Enforce or unenforce 2FA for a specific user (admin)
      operationId: setUserEnforcement
      parameters:
        - name: x-admin-id
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, enforce]
              properties:
                userId:
                  type: string
                enforce:
                  type: boolean
      responses:
        "200":
          description: Per-user enforcement updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  isEnforced:
                    type: boolean
        "400":
          description: Missing userId or enforce
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing x-admin-id header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 2FA not configured for this user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Notifications ─────────────────────────────────────────────────
...